fix(condo): DOMA-4014 refactor code
